fix(tl.py): fix CTL "dj" vs. "j" wrongly swapped, introduced in 3baf54ae36d4b3eaf9ac7d6400b14880b96b9446
CTL "dj" is for transcribing the affricate pronunciation of TL "j",
which occurs in Choân-chiu accents.

CTL "j" is both for transcribing the fricative pronunciation
and for simply transcribing the phoneme
without distinguishing between the fricative and affricate pronunciation.

In 3baf54ae36d4b3eaf9ac7d6400b14880b96b9446,
where the `namedtuple`-based dialect system was introduced,
the transcription definition for Chiang-chiu and Choân-chiu was wrongly swapped.

To address this issue, the transcription definition is swapped back.